Add game state, domain logic, persistence & audit
Introduce core chancellor-sim infrastructure: context/provider, normalisation and persistence layers, domain helpers and scoring, plus project documentation. Key additions: chancellor-sim/src/state/{game-context.tsx,normalisation.ts,persistence.ts}, chancellor-sim/src/domain/budget/apply-changes.ts, chancellor-sim/src/domain/game/scoring.ts, chancellor-sim/src/types.ts, chancellor-sim/src/utils/helpers.ts. Also add audit.md and Frontend design.txt. Updated chancellor-sim/src/game-state.tsx to integrate the new state modules and serialisation/normalisation flow. These changes modularise state handling, prepare domain extraction, and add an audit/design reference for future work.
